requirements.txt
bluepy
python-dotenv
influxdb
flask